requirements.txt
uvicorn[standard]
fastapi
gunicorn
opentelemetry-exporter-otlp-proto-grpc
fastapi==0.95.0
uvicorn==0.20.0
yfinance==0.2.40
elastic-apm==6.22.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-exporter-otlp-proto-http==1.26.0